+++ urbanstats/geometry/shapefiles/shapefiles/judicial.py
@@ -6,25 +6,6 @@ from urbanstats.geometry.shapefiles.shapefile_subset import SelfSubset
 from urbanstats.universe.universe_provider.constants import us_domestic_provider
 
 
-def render_ordinal(x):
-    if x % 100 in {11, 12, 13}:
-        return f"{x}th"
-    if x % 10 == 1:
-        return f"{x}st"
-    if x % 10 == 2:
-        return f"{x}nd"
-    if x % 10 == 3:
-        return f"{x}rd"
-    return f"{x}th"
-
-
-def render_start_and_end(row):
-    start, end = row.start, row.end
-    if start == end:
-        return f"{render_ordinal(start)}"
-    return f"{render_ordinal(start)}-{render_ordinal(end)}"
-
-
 def judicial_districts():
     data = gpd.read_file("named_region_shapefiles/US_District_Court_Jurisdictions.zip")
     data = data[data.STATE.apply(lambda x: us.states.lookup(x) is not None)]
